rename: 파일명 공백 제거
